# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SigmaGrid Fundamentals API",
    "version": "0.1.0",
    "description": "SigmaGrid — Institutional fundamentals API for synthetic-equity perpetuals (SPY-PERP, QQQ-PERP, TSLA-PERP, etc.). Public docs live now — Paid API launching Q1 2026."
  },
  "servers": [
    {
      "url": "https://api.sigmagrid.app",
      "description": "Production (planned, Q1 2026)"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "description": "Returns service health status. This endpoint is public (not x402-gated) and can be used for uptime monitoring and validation checks.",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" },
                "example": {
                  "status": "ok",
                  "service": "sigmagrid-api",
                  "time": "2026-01-15T18:00:00.123456Z"
                }
              }
            }
          }
        }
      }
    },
    "/v1/signals/{ticker}": {
      "get": {
        "summary": "Get consolidated signal snapshot for a ticker",
        "description": "Returns a full signal snapshot for the specified synthetic-equity ticker, including core mandatory fields, extended fields, and optional macro betas.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "Underlying symbol, e.g. SPY, QQQ, TSLA.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Signal snapshot",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignalResponse" },
                "example": {
                  "timestamp": "2026-01-15T18:00:00Z",
                  "ticker": "SPY",
                  "fair_value": 523.41,
                  "fv_upper": 525.10,
                  "fv_lower": 521.80,
                  "drift_1h": 0.32,
                  "drift_overnight": 0.18,
                  "regime": "trend",
                  "vol_forecast_1h": 0.012,
                  "vol_forecast_4h": 0.015,
                  "event_next": "FOMC",
                  "event_impact": "extreme",
                  "event_bias": "risk-off",
                  "premiums": {
                    "hyperliquid": 0.42,
                    "aevo": -0.15
                  },
                  "z_score": 2.1,
                  "reversion_prob": 0.78,
                  "shock_index": 0,
                  "beta_macro": 1.08,
                  "beta_yield": -0.42,
                  "beta_dollar": 0.31,
                  "beta_vol_index": 0.87
                }
              }
            }
          }
        }
      }
    },
    "/v1/premium/{ticker}": {
      "get": {
        "summary": "Get cross-venue basis and mispricing metrics",
        "description": "Returns basis vs fair value and related arbitrage metrics across supported venues.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Premium metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PremiumResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/drift/{ticker}": {
      "get": {
        "summary": "Get directional drift surface for a ticker",
        "description": "Returns short-horizon directional drift fields for the given ticker.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Drift metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DriftResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/regime/{ticker}": {
      "get": {
        "summary": "Get regime and risk sizing information",
        "description": "Returns current regime and shock information for the given ticker, plus optional macro betas.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Regime metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegimeResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/events/{ticker}": {
      "get": {
        "summary": "Get macro event surface for a ticker",
        "description": "Returns the next major macro event, expected impact, and bias for the given ticker.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Event metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/arbitrage/{ticker}": {
      "get": {
        "summary": "Get carry and mean reversion view",
        "description": "Returns fair value, premiums, reversion probability, and stress-aware spreads for the given ticker.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Arbitrage metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArbitrageResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/historical/{ticker}": {
      "get": {
        "summary": "Get historical data for a ticker",
        "description": "Returns historical signal data for the specified ticker for backtesting, calibration, and analysis.",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "description": "Underlying symbol, e.g. SPY, QQQ, TSLA.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HistoricalResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/snapshot": {
      "get": {
        "summary": "Get multi-ticker snapshot",
        "description": "Returns signal snapshots for multiple tickers in a single request.",
        "parameters": [
          {
            "name": "tickers",
            "in": "query",
            "required": true,
            "description": "Comma-separated list of tickers (e.g. SPY,QQQ,TSLA).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Multi-ticker snapshot",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SnapshotResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SignalResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "fair_value": { "type": "number" },
          "fv_upper": { "type": "number" },
          "fv_lower": { "type": "number" },
          "drift_1h": { "type": "number" },
          "drift_overnight": { "type": "number" },
          "regime": {
            "type": "string",
            "description": "trend / chop / HVOL / LVOL"
          },
          "vol_forecast_1h": { "type": "number" },
          "vol_forecast_4h": { "type": "number" },
          "event_next": { "type": "string" },
          "event_impact": {
            "type": "string",
            "description": "Expected impact: low → extreme"
          },
          "event_bias": {
            "type": "string",
            "description": "risk-on / risk-off / neutral"
          },
          "premiums": {
            "type": "object",
            "additionalProperties": { "type": "number" },
            "description": "Map of venue → basis vs fair_value"
          },
          "z_score": { "type": "number" },
          "reversion_prob": { "type": "number" },
          "shock_index": { "type": "number" },
          "beta_macro": { "type": "number" },
          "beta_yield": { "type": "number" },
          "beta_dollar": { "type": "number" },
          "beta_vol_index": { "type": "number" }
        },
        "required": [
          "timestamp",
          "ticker",
          "fair_value",
          "vol_forecast_1h",
          "vol_forecast_4h",
          "regime",
          "drift_1h",
          "drift_overnight",
          "event_next",
          "event_impact",
          "event_bias"
        ]
      },
      "PremiumResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "premiums": {
            "type": "object",
            "additionalProperties": { "type": "number" }
          },
          "z_score": { "type": "number" },
          "reversion_prob": { "type": "number" }
        },
        "required": ["timestamp", "ticker", "premiums"]
      },
      "DriftResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "drift_1h": { "type": "number" },
          "drift_overnight": { "type": "number" },
          "regime": { "type": "string" }
        },
        "required": ["timestamp", "ticker", "drift_1h", "regime"]
      },
      "RegimeResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "regime": { "type": "string" },
          "shock_index": { "type": "number" },
          "beta_macro": { "type": "number" },
          "beta_yield": { "type": "number" },
          "beta_dollar": { "type": "number" },
          "beta_vol_index": { "type": "number" }
        },
        "required": ["timestamp", "ticker", "regime"]
      },
      "EventResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "event_next": { "type": "string" },
          "event_impact": { "type": "string" },
          "event_bias": { "type": "string" }
        },
        "required": ["timestamp", "ticker", "event_next"]
      },
      "ArbitrageResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "ticker": { "type": "string" },
          "fair_value": { "type": "number" },
          "premiums": {
            "type": "object",
            "additionalProperties": { "type": "number" }
          },
          "z_score": { "type": "number" },
          "reversion_prob": { "type": "number" },
          "shock_index": { "type": "number" }
        },
        "required": ["timestamp", "ticker", "fair_value", "premiums"]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Service status, always 'ok' when healthy"
          },
          "service": {
            "type": "string",
            "description": "Service identifier"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "description": "Current server time in ISO 8601 format"
          }
        },
        "required": ["status", "service", "time"]
      }
    }
  }
}